+++ nflow-engine/src/main/java/io/nflow/engine/workflow/definition/StateExecution.java
@@ -189,5 +189,5 @@ public interface StateExecution {
    *
    * @return True if unfinished child workflow instances are found, false otherwise.
    */
-  boolean hasUnfinishedChildren();
+  boolean hasUnfinishedChildWorkflows();
 }