+++ nflow-engine/src/test/java/io/nflow/engine/service/WorkflowInstanceServiceTest.java
@@ -4,6 +4,7 @@ import static io.nflow.engine.workflow.instance.WorkflowInstance.WorkflowInstanc
 import static io.nflow.engine.workflow.instance.WorkflowInstance.WorkflowInstanceStatus.inProgress;
 import static io.nflow.engine.workflow.instance.WorkflowInstanceAction.WorkflowActionType.externalChange;
 import static java.util.Arrays.asList;
+import static java.util.Collections.emptyList;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -12,7 +13,9 @@ import static org.joda.time.DateTimeUtils.currentTimeMillis;
 import static org.joda.time.DateTimeUtils.setCurrentMillisFixed;
 import static org.joda.time.DateTimeUtils.setCurrentMillisSystem;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.doReturn;
@@ -23,6 +26,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.util.Collections;
+import java.util.EnumSet;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -42,6 +46,7 @@ import io.nflow.engine.workflow.definition.AbstractWorkflowDefinition;
 import io.nflow.engine.workflow.definition.WorkflowDefinition;
 import io.nflow.engine.workflow.instance.QueryWorkflowInstances;
 import io.nflow.engine.workflow.instance.WorkflowInstance;
+import io.nflow.engine.workflow.instance.WorkflowInstance.WorkflowInstanceStatus;
 import io.nflow.engine.workflow.instance.WorkflowInstanceAction;
 import io.nflow.engine.workflow.instance.WorkflowInstanceAction.WorkflowActionType;
 
@@ -157,7 +162,7 @@ public class WorkflowInstanceServiceTest extends BaseNflowTest {
 
   @Test
   public void listWorkflowInstances() {
-    List<WorkflowInstance> result  = asList(constructWorkflowInstanceBuilder().build());
+    List<WorkflowInstance> result = asList(constructWorkflowInstanceBuilder().build());
     QueryWorkflowInstances query = mock(QueryWorkflowInstances.class);
     when(workflowInstanceDao.queryWorkflowInstances(query)).thenReturn(result);
     assertEquals(result, service.listWorkflowInstances(query));
@@ -207,4 +212,30 @@ public class WorkflowInstanceServiceTest extends BaseNflowTest {
     verify(workflowDefinitions, never()).getWorkflowDefinition(anyString());
   }
 
+  @Test
+  public void hasUnfinishedChildWorkflowsReturnsFalseWhenInstanceHasNoUnfinishedChildren() {
+    List<WorkflowInstance> result = emptyList();
+    when(workflowInstanceDao.queryWorkflowInstances(queryCapture.capture())).thenReturn(result);
+
+    assertFalse(service.hasUnfinishedChildWorkflows(42));
+    QueryWorkflowInstances query = queryCapture.getValue();
+    assertThat(query.parentWorkflowId, is(42L));
+    EnumSet.complementOf(EnumSet.of(WorkflowInstanceStatus.finished)).stream()
+        .forEach(status -> assertTrue(query.statuses.contains(status)));
+    assertFalse(query.statuses.contains(WorkflowInstanceStatus.finished));
+  }
+
+  @Test
+  public void hasUnfinishedChildWorkflowsReturnsTrueWhenInstanceHasUnfinishedChild() {
+    WorkflowInstance unfinished = constructWorkflowInstanceBuilder().build();
+    List<WorkflowInstance> result = asList(unfinished);
+    when(workflowInstanceDao.queryWorkflowInstances(queryCapture.capture())).thenReturn(result);
+
+    assertTrue(service.hasUnfinishedChildWorkflows(42));
+    QueryWorkflowInstances query = queryCapture.getValue();
+    assertThat(query.parentWorkflowId, is(42L));
+    EnumSet.complementOf(EnumSet.of(WorkflowInstanceStatus.finished)).stream()
+        .forEach(status -> assertTrue(query.statuses.contains(status)));
+    assertFalse(query.statuses.contains(WorkflowInstanceStatus.finished));
+  }
 }